Move ParamFilter enum into new filterutil.go
